Yesod.Default.Util

Description

Various utilities used in the scaffolded site.

Documentation

addStaticContentExternal

Arguments

:: (ByteString -> Either a ByteString)
javascript minifier

-> (ByteString -> String)
hash function to determine file name

-> FilePath
location of static directory. files will be placed within a "tmp" subfolder

-> ([Text] -> Route master)
route constructor, taking a list of pieces

-> Text
filename extension

-> Text
mime type

-> ByteString
file contents

-> HandlerFor master (Maybe (Either Text (Route master, [(Text, Text)])))

An implementation of addStaticContent which stores the contents in an external file. Files are created in the given static folder with names based on a hash of their content. This allows expiration dates to be set far in the future without worry of users receiving stale content.

globFile :: String -> String -> FilePath

expects a file extension for each type, e.g: hamlet lucius julius

widgetFileNoReload :: WidgetFileSettings -> FilePath -> Q Exp

widgetFileReload :: WidgetFileSettings -> FilePath -> Q Exp

data TemplateLanguage

Constructors

TemplateLanguage

Fields

tlRequiresToWidget :: Bool

tlExtension :: String

tlNoReload :: FilePath -> Q Exp

tlReload :: FilePath -> Q Exp

defaultTemplateLanguages :: HamletSettings -> [TemplateLanguage]

data WidgetFileSettings

wfsLanguages :: WidgetFileSettings -> HamletSettings -> [TemplateLanguage]

wfsHamletSettings :: WidgetFileSettings -> HamletSettings